+++ python/dmlab_module.c
@@ -1,4 +1,4 @@
-// Copyright (C) 2016-2017 Google Inc.
+// Copyright (C) 2016-2018 Google Inc.
 //
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -40,19 +40,19 @@
 // Glue code to make the code work with both Python 2.7 and Python 3.
 //
 #if PY_MAJOR_VERSION >= 3
-#  define PY_INIT_SIGNATURE(name) PyInit_ ## name
-#  define PY_RETURN_SUCCESS(x) return x
-#  define PY_RETURN_FAILURE    return NULL
 #  define PyInt_FromLong PyLong_FromLong
 #  define PyInt_AsLong PyLong_AsLong
 #  define PyInt_Check PyLong_Check
-#else  // PY_MAJOR_VERSION >= 3
-#  define PY_INIT_SIGNATURE(name) init ## name
-#  define PY_RETURN_SUCCESS(x) return
-#  define PY_RETURN_FAILURE    return
 #endif  // PY_MAJOR_VERSION >= 3
 
-static char runfiles_path[4096];  // set in initdeepmind_lab() below
+// In both Py2 and Py3, the LabModuleState object(s) that we will be using are
+// zero-initialized.
+typedef struct {
+  char runfiles_path[4096];  // Populated during module initialization below.
+  PyTypeObject lab_type;
+} LabModuleState;
+
+static LabModuleState* get_module_state(PyObject* module);  // defined below
 
 typedef struct {
   PyObject_HEAD
@@ -165,8 +165,18 @@ static int Lab_init(PyObject* pself, PyObject* args, PyObject* kwds) {
     return -1;
   }
   {
+#if PY_MAJOR_VERSION >= 3
+    PyObject* module = PyImport_AddModule("deepmind_lab");
+    if (module == NULL) {
+      PyErr_SetString(PyExc_RuntimeError, "deepmind_lab module not loaded");
+      return -1;
+    }
+#else  // PY_MAJOR_VERSION >= 3
+    PyObject* module = NULL;
+#endif  // PY_MAJOR_VERSION >= 3
+
     DeepMindLabLaunchParams params = {};
-    params.runfiles_path = runfiles_path;
+    params.runfiles_path = get_module_state(module)->runfiles_path;
     params.renderer = DeepMindLabRenderer_Software;
     if (renderer != NULL && renderer[0] != '\0') {
       if (strcmp(renderer, "hardware") == 0) {
@@ -721,7 +731,7 @@ static PyTypeObject deepmind_lab_LabType = {
 };
 
 static PyObject* module_runfiles_path(PyObject* self, PyObject* no_arg) {
-  return Py_BuildValue("s", runfiles_path);
+  return Py_BuildValue("s", get_module_state(self)->runfiles_path);
 }
 
 static PyObject* module_set_runfiles_path(PyObject* self, PyObject* args) {
@@ -730,8 +740,9 @@ static PyObject* module_set_runfiles_path(PyObject* self, PyObject* args) {
     return NULL;
   }
 
-  if (strlen(new_path) < sizeof(runfiles_path)) {
-    strcpy(runfiles_path, new_path);
+  LabModuleState* module_state = get_module_state(self);
+  if (strlen(new_path) < sizeof(module_state->runfiles_path)) {
+    strcpy(module_state->runfiles_path, new_path);
   } else {
     PyErr_SetString(PyExc_RuntimeError, "Runfiles directory name too long!");
     return NULL;
@@ -754,66 +765,57 @@ static PyMethodDef module_methods[] = {
     {NULL, NULL, 0, NULL} /* sentinel */
 };
 
+
+////////////////////////////////////////////////////////////////////////////////
+// Module initialization:
+//
+// This part looks quite different in Python 2 and Python 3.
+////////////////////////////////////////////////////////////////////////////////
+
+static int load_module_impl(PyObject* module, LabModuleState* state) {
 #if PY_MAJOR_VERSION >= 3
-static PyModuleDef module_def = {
-    PyModuleDef_HEAD_INIT,
-    "deepmind_lab",             /* m_name */
-    "DeepMind Lab API module",  /* m_doc */
-    -1,                         /* m_size */
-    module_methods,             /* m_methods */
-    NULL,                       /* m_reload */
-    NULL,                       /* m_traverse */
-    NULL,                       /* m_clear */
-    NULL,                       /* m_free */
-};
-#endif  // PY_MAJOR_VERSION >= 3
-
-PyMODINIT_FUNC PY_INIT_SIGNATURE(deepmind_lab)(void) {
-  PyObject* m;
-
-  if (PyType_Ready(&deepmind_lab_LabType) < 0) PY_RETURN_FAILURE;
-
-#if PY_MAJOR_VERSION >= 3
-  m = PyModule_Create(&module_def);
+  PyTypeObject* lab_type = &state->lab_type;
+  memcpy(lab_type, &deepmind_lab_LabType, sizeof(PyTypeObject));
 #else  // PY_MAJOR_VERSION >= 3
-  m = Py_InitModule3("deepmind_lab", module_methods, "DeepMind Lab API module");
+  PyTypeObject* lab_type = &deepmind_lab_LabType;
 #endif  // PY_MAJOR_VERSION >= 3
-
-  Py_INCREF(&deepmind_lab_LabType);
-  PyModule_AddObject(m, "Lab", (PyObject*)&deepmind_lab_LabType);
+  if (PyType_Ready(lab_type) < 0) return -1;
+  Py_INCREF(lab_type);
+  PyModule_AddObject(module, "Lab", (PyObject*)lab_type);
 
 #ifdef DEEPMIND_LAB_MODULE_RUNFILES_DIR
-  PyObject *v = PyObject_GetAttrString(m, "__file__");
+  PyObject *v = PyObject_GetAttrString(module, "__file__");
 #if PY_MAJOR_VERSION >= 3
-  if (v && PyBytes_Check(v)) {
-    const char* file = PyBytes_AsString(v);
+  if (v && PyUnicode_Check(v)) {
+    const char* file = PyBytes_AsString(PyUnicode_EncodeFSDefault(v));
 #else  // PY_MAJOR_VERSION >= 3
   if (v && PyString_Check(v)) {
     const char* file = PyString_AsString(v);
 #endif  // PY_MAJOR_VERSION >= 3
-    if (strlen(file) < sizeof(runfiles_path)) {
-      strcpy(runfiles_path, file);
+    if (strlen(file) < sizeof(state->runfiles_path)) {
+      strcpy(state->runfiles_path, file);
     } else {
       PyErr_SetString(PyExc_RuntimeError, "Runfiles directory name too long!");
-      PY_RETURN_FAILURE;
+      return -1;
     }
 
-    char* last_slash = strrchr(runfiles_path, '/');
+    char* last_slash = strrchr(state->runfiles_path, '/');
     if (last_slash != NULL) {
       *last_slash = '\0';
     } else {
       PyErr_SetString(PyExc_RuntimeError,
                       "Unable to determine runfiles directory!");
-      PY_RETURN_FAILURE;
+      return -1;
     }
   } else {
-    strcpy(runfiles_path, ".");
+    fprintf(stderr, "Failed to get __file__ attribute.\n");
+    PyErr_Clear();
+    strcpy(state->runfiles_path, ".");
   }
-#else
+#else  // DEEPMIND_LAB_MODULE_RUNFILES_DIR
 #if PY_MAJOR_VERSION >= 3
   PyObject* u = PyUnicode_FromWideChar(Py_GetProgramFullPath(), -1);
-  if (u == NULL) PY_RETURN_FAILURE;
-
+  if (u == NULL) return -1;
   PyObject* p = PyUnicode_EncodeFSDefault(u);
   const char* s = PyBytes_AsString(p);
   size_t n = PyBytes_Size(p);
@@ -822,18 +824,89 @@ PyMODINIT_FUNC PY_INIT_SIGNATURE(deepmind_lab)(void) {
   size_t n = strlen(s);
 #endif  // PY_MAJOR_VERSION >= 3
   static const char kRunfiles[] = ".runfiles/org_deepmind_lab";
-  if (n + strlen(kRunfiles) < sizeof(runfiles_path)) {
-    strcpy(runfiles_path, s);
-    strcat(runfiles_path, kRunfiles);
+  LabModuleState* module_state = get_module_state(module);
+  if (n + strlen(kRunfiles) < sizeof(module_state->runfiles_path)) {
+    strcpy(module_state->runfiles_path, s);
+    strcat(module_state->runfiles_path, kRunfiles);
   } else {
     PyErr_SetString(PyExc_RuntimeError, "Runfiles directory name too long!");
-    PY_RETURN_FAILURE;
+    return -1;
   }
-#endif
+#endif  // DEEPMIND_LAB_MODULE_RUNFILES_DIR
 
   srand(time(NULL));
-
-  import_array();
-
-  PY_RETURN_SUCCESS(m);
+  return 0;
 }
+
+#if PY_MAJOR_VERSION >= 3
+
+static PyObject* dmlab_create_mod(PyObject* spec, PyModuleDef* def) {
+  PyObject* modname = PyObject_GetAttrString(spec, "name");
+  if (modname == NULL) return NULL;
+
+  PyObject* module = PyModule_NewObject(modname);
+  Py_DECREF(modname);
+  if (module == NULL) return NULL;
+
+  PyObject* moddict = PyModule_GetDict(module);
+  PyObject* v = PyObject_GetAttrString(spec, "origin");
+  if (v == NULL) return NULL;
+  int res = PyDict_SetItemString(moddict, "__file__", v);
+  Py_DECREF(v);
+  if (res != 0) return NULL;
+
+  return module;
+}
+
+static int dmlab_exec_mod(PyObject* module) {
+  if (load_module_impl(module, PyModule_GetState(module)) == -1) {
+    return -1;
+  } else {
+    import_array1(-1);
+    return 0;
+  }
+}
+
+static PyModuleDef_Slot module_slots[] = {
+  {Py_mod_create, dmlab_create_mod},
+  {Py_mod_exec, dmlab_exec_mod},
+  {0, NULL},
+};
+
+static PyModuleDef module_def = {
+    PyModuleDef_HEAD_INIT,
+    "deepmind_lab",             /* m_name */
+    "DeepMind Lab API module",  /* m_doc */
+    sizeof(LabModuleState),     /* m_size */
+    module_methods,             /* m_methods */
+    module_slots,               /* m_slots */
+    NULL,                       /* m_traverse */
+    NULL,                       /* m_clear */
+    NULL,                       /* m_free */
+};
+
+PyMODINIT_FUNC PyInit_deepmind_lab(void) {
+  return PyModuleDef_Init(&module_def);
+}
+
+static LabModuleState* get_module_state(PyObject* module) {
+  return PyModule_GetState(module);
+}
+
+#else  // PY_MAJOR_VERSION >= 3
+
+static LabModuleState singleton_mod_state;
+static LabModuleState* get_module_state(PyObject* module) {
+  return &singleton_mod_state;
+}
+
+PyMODINIT_FUNC initdeepmind_lab(void) {
+  PyObject* module =
+      Py_InitModule3("deepmind_lab", module_methods, "DeepMind Lab API module");
+
+  if (module != NULL && load_module_impl(module, &singleton_mod_state) == 0) {
+    import_array();
+  }
+}
+
+#endif  // PY_MAJOR_VERSION >= 3